flip t before processing it
The issue was caused because np.unique uses the first occurrence for duplicates. Break data at the beginning of the replay would then be replacing the data of the actual replay, causing jumps to incorrect coordinates. Flipping the t array before processing it makes it so np.unique chooses the actual frames instead of the ones in the break.